move getSession to spotify utils dir
