+++ src/spotify/utils/getSession.ts
@@ -1,7 +1,7 @@
 import { type AccessToken } from '@spotify/web-api-ts-sdk';
 import { cookies } from 'next/headers';
 
-import { SPOTIFY_AUTH_COOKIE } from './constants';
+import { SPOTIFY_AUTH_COOKIE } from '../constants';
 
 export const tryGetAuthSession = (): AccessToken | null => {
   const authSessionString = cookies().get(SPOTIFY_AUTH_COOKIE)?.value;
